fixed problems with schemas, new step in common steps and documented functions
