+++ main/utils/json_model.py
@@ -1,0 +1,6 @@
+class JsonModel():
+    def __init__(self):
+        self.json_data = {}
+
+    def build_json(self, key, value):
+        self.json_data[key] = value
