+++ tests/api/features/boards.feature
@@ -2,16 +2,19 @@ Feature: Trello Boards API
     As a user I want to use Boards API and perform Get,Post And Delete operations
    
    Scenario: Get all information of board 
+        Given I am a trello user
         When I send a "GET" request to "/boards/{board}"
         Then the status code is "200"
         And I receive a response with the "board" schema
 
     Scenario: Get members of board 
+        Given I am a trello user
         When I send a "GET" request to "/boards/{board}/members"
         Then the status code is "200"
-        And verify there is at least 1 member
+        And I receive a list with at least "1" "{member}"
     
     Scenario: Updating information of a Board
+        Given I am a trello user
         When I send a "PUT" request to "/boards/{board}"
             | Key         | Value                      |
             | name        | testingnicolasv            |
@@ -23,7 +26,7 @@ Feature: Trello Boards API
             | desc        | a valid data table         | 
 
 
-     Scenario Outline: CREATE and DELETE a board 
+    Scenario Outline: CREATE and DELETE a board 
         Given the user creates a board with "<field>" to be "<value>"
         Then the status code is "200"
         And I receive a response with the "board" schema
@@ -34,3 +37,10 @@ Feature: Trello Boards API
         | field | value     |
         | name  | kiara     |
         | name  | jose      |
+        | name  | Nicolas   |
+        | name  | keila     |
+        | name  | Vallejo   |
+        | name  | pabon     |
+
+    
+
