+++ tests/api/features/steps/boards_steps.py
@@ -2,21 +2,16 @@ from logging import exception
 from behave import given, when, then, step
 from jsonschema import validate
 from main.core.rest.request_manager import RequestManager
-
 from main.trello.api.boards_manager import BoardsManager
-
-
+from main.utils.json_model import JsonModel
 request_manager = RequestManager()
 boards_manager = BoardsManager()
 
-
-
-@step('verify there is at least 1 member')
-def step_impl(context):    
-    assert len(context.response), f"cound't retrieve members from board "
-
 @step('verify the new object contains the following info')
-def step_impl(context):    
+def step_impl(context): 
+    """ 
+    Construct a Json with the table of the .features and compare it with the response frome the API 
+    """   
     model = JsonModel()
     for row in context.table:
         model.build_json(row['Key'], row['Value'])
@@ -24,21 +19,31 @@ def step_impl(context):
     for key, value in model.json_data.items():
         actual_value =  context.response[key]
         assert value == actual_value, f"Expected: {value}, Actual: {actual_value} "
-
-
+        
 @step ('the user creates a board with "{field}" to be "{value}"')
 def step_impl(context, field, value ): 
+    """ 
+    Construct a Json with the table of the .features in order to send that as the payload and then creates a board using the payload
+
+    :param field:   String  string with the name of the field that will be filled when creating the board
+    :param value:   String  string with the infomation that would be placed in the field
+    """  
     context.model = JsonModel()
     context.model.build_json(field, value)
-
     name = context.model.json_data["name"] 
     status_code, response =  boards_manager.create_board(name )
     context.status_code = status_code
     context.response = response
 
-
 @step ('the user deletes the board "{name}"')
 def step_impl(context, name): 
+    """ 
+    recives a name of a board and compares it with the board in the context, if they are the same, it uses the board manager 
+    to delete the board
+
+    :param name:   String  string with the name of the board that will be deleted 
+    """
+
     if (context.response["name"] == name):
         context.status_code, context.response = boards_manager.delete_board(context.response["id"])
 
@@ -48,9 +53,4 @@ def step_impl(context, name):
 
 
 
-class JsonModel():
-    def __init__(self):
-        self.json_data = {}
 
-    def build_json(self, key, value):
-        self.json_data[key] = value
